Move qrexec_popen, qrexec_payload, and qrexec_repoquery into templates module
